スコアとゲージの計算を実装

+++ TaiCombo.Plugin/src/Chart/Course.cs
@@ -4,13 +4,13 @@ namespace TaiCombo.Plugin.Chart;
 
 public class Course
 {
-    public Course(float level, List<IChip> chips)
+    public Course(int level, List<IChip> chips)
     {
         Level = level;
         Chips = chips;
     }
 
-    public float Level { get; set; }
+    public int Level { get; set; }
 
     public List<IChip> Chips { get; set; }
 }
